warnings cleanup
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9086 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDRenderer.cxx
@@ -319,7 +319,7 @@ void			HUDRenderer::setTeamHasHighScore(bool hasHigh)
 
 void			HUDRenderer::setHeading(float angle)
 {
-  heading = 90.0f - 180.0f * angle / M_PI;
+  heading = (float)(90.0 - 180.0 * angle / M_PI);
   while (heading < 0.0f) heading += 360.0f;
   while (heading >= 360.0f) heading -= 360.0f;
 }
